# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "System Rejestracji Pacjenta Stomatologicznego API",
      "description": "API dla systemu rejestracji pacjenta stomatologicznego",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/api",
        "description": "Development server"
      }
    ],
    "components": {
      "securitySchemes": {
        "jwt": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "paths": {
      "/auth/register/pacjent": {
        "post": {
          "tags": ["Autoryzacja"],
          "summary": "Rejestracja pacjenta",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "imie": {"type": "string"},
                    "nazwisko": {"type": "string"},
                    "email": {"type": "string"},
                    "haslo": {"type": "string"},
                    "telefon": {"type": "string"}
                  },
                  "required": ["imie", "nazwisko", "email", "haslo", "telefon"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Konto pacjenta utworzone pomyślnie"
            },
            "400": {
              "description": "Błędne dane lub adres email jest już używany"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "tags": ["Autoryzacja"],
          "summary": "Logowanie użytkownika",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {"type": "string"},
                    "haslo": {"type": "string"},
                    "rola": {"type": "string", "enum": ["pacjent", "lekarz", "admin"]}
                  },
                  "required": ["email", "haslo", "rola"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Zalogowano pomyślnie"
            },
            "401": {
              "description": "Nieprawidłowy email lub hasło"
            }
          }
        }
      },
...
    }
  }